requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
cloudinary==1.41.0
colorama==0.4.6
comtypes==1.4.8
distro==1.9.0
edge-tts==7.0.0
Flask==3.0.0
frozenlist==1.5.0
groq==0.13.1
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.2.1
pydantic==2.10.3
pydantic_core==2.27.1
pygame==2.6.1
pypiwin32==223
python-dotenv==1.0.0
pyttsx3==2.98
pywin32==308
requests==2.31.0
six==1.17.0
sniffio==1.3.1
srt==3.5.3
tabulate==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.1
yarl==1.18.3
